Add unit test for FEPSPlumeRise.compute
This test monkeypatches away the running of feps_weather and
feps_plumerise.  Given that these two executables do the actual
plumerise computation, this test basically just verifies that all
input files are written correctly, and that the output plume file is
read correctly.  With the assumption that the FEPS executables function
correctly, this unit test can be viewed as a de facto regression test.